Fix handling of tts:false metadata
ConversationsController#update actually only checked for presence of
meta_data['tts'] and in case it was not existing, used true as a default.
But with the || operator for test/initialization of this value,
of course an eventually existing false setting was also overridden to true.

As we have multiple passes over JSON decode and Ruby hashes, implement a
true? method to safeguard against String/Boolean conversion for boolean
tests.

Changed one test fixture to meta_data 'tts:false' so that we have both
cases handled.

Signed-off-by: Daniel Schnell <dschnell@grammatek.com>