Add GET /info
GET /info returns the MOTD(message of the day) and all supported languages
of the dialog system.
It tries to communicate to Rasa via the configured `rasa_intent` to
get the MOTD from there and in case this doesn't succeed, provides a
default implementation according to what has been configured inside
`config/masdif.yml`.

- implement Rasa REST API backend communication methods model_parse and
  trigger_intent
- adds configuration options to config/masdif.yml for languages and motd
- implement info_controller, configure routes & add test cases for endpoints
  and configuration
- add documentation to README.md

Signed-off-by: Daniel Schnell <dschnell@grammatek.com>