ConversationsController: Add meta-data entry :asr_generated
Add new meta-data property :asr_generated. This will be set by the
web widget in case the sent text was generated via ASR.
Permit this paramter in the conversation_params checker. The flag is saved
into the database like any other meta-data that is sent to the controller.

Also add paramater to fixtures for existing tests to exercise the paramter.

Signed-off-by: Daniel Schnell <dschnell@grammatek.com>