Implement /feedback handling for conversation
The client can send feedback for a message response to the endpoint
/conversation/<id>. This feedback is saved into the DB. It can be
configured, if the feedback is forwarded to the dialog system or if
it should be just saved into DB. If feedback is provided, the message
payload that the client sends needs to contain a message_id of a previously
received message.
If the message_id is provided, but the body of the request doesn't contain
the required format /feedback{"value":"someval"}, an error is returned.
Errors are also returned in case the message_id isn't present, but the
feedback is sent.

- add  test cases for good and bad feedback paths, also add test fixtures
- fix mandatory field checks for #check_bot_response and change a button
  initiating message
- move feedback column from model Conversation to model Message, add
  migration and update fixtures + code
- add default configuration for Rails.application.config.masdif[:feedback]
- Fix forwarding reply text to Rasa. So far we were saving the complete Rasa
  response inside the reply field of a message instead of just the json reply
  to the client. Therefore, it was difficult to extract just the relevant
  text from it. Now we save the JSON reply that is returned to the client
  and parse it to extract the 'text' fields of it.
- Update README.md

Signed-off-by: Daniel Schnell <dschnell@grammatek.com>