Add basic admin interface via ActiveAdmin
Provide standard configuration for ActiveAdmin and make available
the resources conversations, messages.

So far very basic and raw, but already usable for cleaning up and
debugging.

Moreover change the database type of message#reply from string to jsonb
and convert existing data of that column to jsonb inside the db migration.
In the same context, fix generation of the message reply JSON data inside
conversation#create

Signed-off-by: Daniel Schnell <dschnell@grammatek.com>