Process custom actions from the dialog system.
If a key value 'custom' is present inside the responses array, the value
of that key is an action reply containing JSON. Save the action reply to
the message and remove the custom element from the reply attribute.

Also do this for existing db columns and add an appropriate migration.

Signed-off-by: Daniel Schnell <dschnell@grammatek.com>