Update TTS_ATTACHMENT_TIMEOUT to 14 days inside .env.example
The environmemnt file .env is used in Masdif to set environment
variables. These are e.g. evaluated inside `config/masdif.yml` to fill in
configuration values and if a particualr environemnt variable is missing,
an alternative default configuration value is used.

As we are using .env.example for deployment on https://municipality.sdifi.com,
we need to adapt the environment file values for the setting TTS_ATTACHMENT_TIMEOUT.

Remove parsing of the .env and .enf.deploly files, as all of the parsing
of .env has already happened at that time.

Add clarification for precedence of env variables to README.md.

Signed-off-by: Daniel Schnell <dschnell@grammatek.com>