docker-compose.yml: fix local disk storage for docker-compose
In case of a container restart, all audio attachments have been lost,
because the directory ./storage was not placed inside a docker volume.

OTOH the database still saved the metadata of the attachments, so that
one could still see audio attachments inside the Admin interface, but those
cannot be played because of the missing raw files.

This commit adds the docker volume 'masdif-storage' to the container which
will be automatically created in case it did not already existed before.

Signed-off-by: Daniel Schnell <dschnell@grammatek.com>