Bump activeadmin to 3.0.0
The dependent Gem Ransacker has been upped to 4.0 making an incompatible
change necessary to the models.

This MR adds the necessary changes for explicitely stating the attributes
that may be searchable by Ransack. We allow all attributes to be searchable
for most models, but for the user model the searchable attributes are
explictely enumerated, excluding e.g. password.

Signed-off-by: Daniel Schnell <dschnell@grammatek.com>